projects/AICouncil/services/council/requirements.txt
fastapi==0.95.0
uvicorn==0.17.0
pydantic==2.0.0
python-dotenv==1.0.0
